requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-huggingface
chromadb
pymupdf
pillow
python-dotenv
sentence-transformers
